requirements.txt
# Core dependencies
fastapi>=0.104.1,<1.0.0
uvicorn>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# OpenAI and LLM
tiktoken>=0.5.2,<1.0.0
openai>=1.3.7,<2.0.0
langchain>=0.0.350,<1.0.0
langchain-openai>=0.0.2,<1.0.0

# Vector database and embeddings
faiss-cpu>=1.7.4,<2.0.0
numpy>=1.24.3,<2.0.0

# PDF Processing
PyPDF2>=3.0.1,<4.0.0
pdfplumber>=0.10.3,<1.0.0

# Data processing
pandas>=2.0.3,<3.0.0

# Authentication (optional)
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Pydantic
pydantic>=2.5.0,<3.0.0

# Build dependencies (included for reference, not installed by default)
# setuptools>=65.5.1
# wheel>=0.40.0
# Cython>=0.29.33